Moved EmailAndNameWorker to java/com/flowcrypt/email/jetpack/workmanager.| #1717
